+++ CHANGELOG.md
@@ -43,6 +43,9 @@ Miscellaneous:
   view by appending `&help=true`. This will be added automatically if a help
   dialog is open during URL creation.
 
+- Connectivity widget: the new "List links' button will open a Connector List
+  with all links of the selected neurons.
+
 - CLI importer: recompute materializations (edges, connectors) explicitly only
   for imported data, which improves performance in typical scenarios (import is
   small compared to existing data). If the old behavior of recomputing